+++ NEWS.md
@@ -8,6 +8,7 @@ taxlist 0.2.1
 * Function `replace_idx()` setting by default `idx1=x`.
 * Functions `replace_idx()` and `replace_na()` setting by default `idx2=idx1`.
 * Special characters corrected in data set *Cyperus*.
+* Validation allowing taxa without rank but parents.
 
 taxlist 0.2.0
 =============