+++ NEWS.md
@@ -4,6 +4,7 @@ taxlist 0.2.1
 ### New Features
 
 * New function `indented_list()` to print taxonomic ranks in indented lists.
+* Slot **taxonViews** in `taxlist` objects extended to class `lib_df`.
 
 ### Improvements